smart-contract-audit/requirements.txt
Flask==3.1.0
flask-cors==6.0.0
google-generativeai==0.8.5
openai==1.84.0
requests==2.32.3
python-dotenv==1.1.0
anthropic==0.40.0